LibSpiro version 0.3
Collective fixes from tagpoint v0.2 (20130930) up to now (20150131).
Most fixes are for setting-up libspiro to be seen as a package, and to
solve some bugs listed on various distro sites. Also added two functions
similar to original TaggedSpiroCPsToBezier() and SpiroCPsToBezier() for
use by developers that need functions that require a void/nothing return.