Reduced multi-pthread test to 100xN tests, add more AC_CHECK_HEADERs
15 minute time-out happened with debian hurd i386, therefore reduced
multi-pthread tests down to 100xN tests instead of 3000 tests.
Added --enable-test-a-lot so that developers, can run a larger batch.

FreeBSD failed to compile due to missing gettimeofday which was present
in version0.3. Edited test files so that tests can be run on systems
that have gettimeofday() and systems running the older ftime().