Allow Spiros to be scaled and/or shifted. Scaling bug fixed.
A scaling bug was found when creating call-testm.c. spiro.c:integrate has
comment mentioning math is for {-0.5..+0.5} calculations. This allows a
spiro to be shifted and/or resized (and produce same iteration results).
This bug fix breaks backwards compatibilty with large, tiny, or shifted
spiros, but for least (expected) 'font' breakage spiro calculations are
done for spiros in (expected) range of {0..1000} instead of {-0.5..+0.5}.
Tests 14 and 15 test large and tiny spiros, with same iteration # results,
and call-testm.c now allowed to scale and/or shift as originally intended.