Add ARC output - this maybe useful for drawing programs.
Also experiment with building spiros with fewer points.
path4[] might be built with fewer points for a cubic curve, allowed
bend to be as large as 90deg, but we may need to reduce this value.

Fix loop & 'z' results to use original src[0] x,y values to avoid
rounding possibe rounding errors.