Add spiroreverse input - this maybe useful for some programs.
During one discussion in FontForge concerning the merging of spiros,
it was mentioned that there may be a need to reverse the direction
of the spiro. This repopulates the input source path with a reversed
version which you can then re-use again, or keep for a 'new' path.

The added tests are all visual, and checked using verbose mode.