maintain backwards compatibility with older interface
Users running older software will only have access to spiroentry
points up to 1, but not to spiroentry point 2. Maintain backwards
compatibility for these older access points.

Users with access to the newer spiroentry points will default to
the better math calculations that can be scaled or moved, and also
have the choice to be RETRO if necessary for older data that needs
to use the older calculations.

Calculations are approximately the same between older and newer
methods for x/Y data within 0...1000, but not 100%, so the user
needs an option RETRO for these occasions.

tests 14 and up, use the new scaled and moveable calculations, and
older tests remain using older tests without scaling or moveability.