fix set_di_to_x1y1() box boundaries relative to curve size
Boundary window is only 0.0005 of curve, but the curve can
also be 1e+40 or 1e-40 in size, so we need to adjust size.
Then return our original source xy points (if we have it).
Prior window was scaling wrong as deltaX or deltaY from 0.

and also visually check a few options for the 2 functions.