LibSpiro 20221101 - Version 1.3.0 Fixed two bugs in spiro.c 1. Removed test code "if (xmax >= ymax) dm[0] = xmax; else dm[0] = ymax;" This would be seen as problematic if drawing in the -x, -y area. (This was test-code that I forgot to remove before making a release). 2. Moved NULL test to correct location (SPIRO_INTERNAL_BEZCTX=0). Fixed minor numerical bug in call-test21.c (This bug is not important to anyone, except maybe/not maintainers).
Little bit of a bonus...
Solved call-test10 and call-test11 by searching for the first available
'c' or 'o'. This is more of a kludge versus a true fix since a spiro
curve will need to contain one or more 'v', 'c' or 'o' somewhere.
call-test10 is basically the same curve as call-test7, but starts with
'ah', while call-test11 is a curve starting with '[]'. Basically KISS.